Remove RuleSet
